+++ porky_v1/src/main/java/com/aydsii/porky/controlador.java
@@ -16,30 +16,30 @@ public class Controlador {
         model.put("template","template/indice.vsl");
         return new VelocityTemplateEngine().render(new ModelAndView(model, "template/layout.vsl"));
     };
+
     @SuppressWarnings("rawtypes")
     public static Route allProductos = (Request request, Response response) -> {
         HashMap model = new HashMap();
-        List<Producto> RES = Producto.buscarProducto();
+        List<Producto> RES = ProductoDAO.buscarProducto();
         //System.out.print(RES);
         model.put("RES", RES);
         model.put("template","template/carta.vsl");
         return new VelocityTemplateEngine().render(new ModelAndView(model, "template/layout.vsl"));
     };
+
     @SuppressWarnings("rawtypes")
     public static Route moreInfo = (Request request, Response response) -> {
         HashMap model = new HashMap();
-        List<Producto> RES = Producto.masInformacion(1);
-        System.out.println(RES);
+        List<Producto> RES = ProductoDAO.masInformacion(Integer.valueOf(request.queryParams("id")));
+        //System.out.println(RES);
         model.put("RES", RES);
         model.put("template","template/infoproducto.vsl");
         return new VelocityTemplateEngine().render(new ModelAndView(model, "template/layout.vsl"));
     };
+    
     @SuppressWarnings("rawtypes")
     public static Route admin = (Request request, Response response) -> {
         HashMap model = new HashMap();
-        List<Producto> RES = Producto.masInformacion(1);
-        System.out.println(RES);
-        model.put("RES", RES);
         model.put("template","template/infoproducto.vsl");
         return new VelocityTemplateEngine().render(new ModelAndView(model, "template/layout.vsl"));
     };
